// src/components/ui/Skeleton.d.ts
import React from 'react';
interface SkeletonProps {
    variant?: 'text' | 'circle' | 'rect';
    width?: string | number;
    height?: string | number;
    count?: number;
    className?: string;
}
export declare const Skeleton: React.FC<SkeletonProps>;
export {};
//# sourceMappingURL=Skeleton.d.ts.map